+++ src/main/java/dnd/dnd10_backend/common/domain/CommonResponse.java
@@ -23,16 +23,14 @@ import lombok.Setter;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class CommonResponse<T> {
+public class CommonResponse {
     ResponseStatus status;
     int code;
     String message;
-    T data;
 
-    public CommonResponse(ResponseStatus status, CodeStatus codeStatus,T data){
+    public CommonResponse(ResponseStatus status, CodeStatus codeStatus){
         this.status = status;
         this.code = codeStatus.getCode();
         this.message = codeStatus.getMessage();
-        this.data = data;
     }
 }
