+++ src/main/java/dnd/dnd10_backend/user/domain/enums/Role.java
@@ -1,5 +1,10 @@
 package dnd.dnd10_backend.user.domain.enums;
 
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonValue;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
 /**
  * 패키지명 dnd.dnd10_backend.user.domain.enums
  * 클래스명 Role
@@ -11,7 +16,26 @@ package dnd.dnd10_backend.user.domain.enums;
  * [수정내용]
  * 예시) [2022-09-17] 주석추가 - 원지윤
  */
-
+@Getter
+@AllArgsConstructor
 public enum Role {
-    WORKER, MANAGER
+    WORKER("WORKER"),
+    MANAGER("MANAGER");
+
+    private final String value;
+
+    @JsonCreator
+    public static Role from(String sub) {
+        for (Role role : Role.values()) {
+            if (role.getValue().equals(sub)) {
+                return role;
+            }
+        }
+        return null;
+    }
+
+    @JsonValue
+    public String getValue() {
+        return value;
+    }
 }
