+++ src/main/java/dnd/dnd10_backend/user/dto/response/UserResponseDto.java
@@ -1,5 +1,10 @@
 package dnd.dnd10_backend.user.dto.response;
 
+import dnd.dnd10_backend.user.domain.User;
+import dnd.dnd10_backend.user.domain.enums.Role;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 /**
@@ -13,10 +18,25 @@ import lombok.Setter;
  * [수정내용]
  * 예시) [2022-09-17] 주석추가 - 원지윤
  */
-@Setter
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
 public class UserResponseDto {
+    private Long userCode;
     private String kakaoNickname;
     private String kakaoEmail;
+    private Role role;
     private String workTime;
     private String workPlace;
+
+    public static UserResponseDto of(User user){
+        return new UserResponseDto(
+                user.getUserCode(),
+                user.getKakaoNickname(),
+                user.getKakaoEmail(),
+                user.getRole(),
+                user.getWorkTime(),
+                user.getWorkPlace()
+        );
+    }
 }
